add db recovery and ut

+++ internal/kv/file/option.go
@@ -1,7 +1,7 @@
 package file
 
 type Option struct {
-	FID      uint32
+	FID      uint64
 	FileName string
 	Dir      string
 	Path     string
